Fix for: error: 'when' expression must be exhaustive, add necessary 'null' branch or 'else' branch instead
AOSP DeskClock contains multiple warnings that become errors when updating to Kotlin 1.7. One such error is that 'when' expressions must be exhaustive (even if all cases are functionally covered). A simple fix for this is to simply add an empty null branch (i.e. null -> { }) at the end of the when statement.

https://android.googlesource.com/platform/packages/apps/DeskClock/+/2d796a4ee72adf6d2016f7650c50b8eb6ec0c8e7%5E%21/#F0