Fixed warning: 'invalidate(Int, Int, Int, Int): Unit' is deprecated. Deprecated in Java
From https://developer.android.com/reference/android/view/View#invalidate() - This method was deprecated in API level 28. The switch to hardware accelerated rendering in API 14 reduced the importance of the dirty rectangle. In API 21 the given rectangle is ignored entirely in favor of an internally-calculated area instead. Because of this, clients are encouraged to just call invalidate().

Replaced instances of invalidate(Int, Int, Int, Int) with a call to invalidate(). Commented out invalidate(Float, Float, Float) implementation as it is now unused (it could also be removed).

https://android.googlesource.com/platform/packages/apps/DeskClock/+/954ff9ebece1441f43723fe64154a592b5701397%5E%21/#F0