Update targetSDKlevel to 30
When the targetSDKlevel is updated, changes are required
in setting notifications for timers.

Updated the logic to build, set and trigger notificaitons
for timers to prevent issues that were reported in bugs
like b/199822689

This is the first in a series of patches that will try to
bump the targetSDKlevel to 33

Multiple manual tests run to check functionality.

https://android.googlesource.com/platform/packages/apps/DeskClock/+/a899ae827b670e462b7c488fcbb0095059b9adf6%5E%21/#F0